feat(teax): 新增审计日志页面与项目设置管理功能
- 新增审计日志页面：`/admin/audit-logs`，展示用户操作记录、资源类型、IP 地址等信息
- 新增审计日志导航：后台管理侧边栏添加"审计日志"菜单项
- 新增操作类型映射：支持项目创建/删除、权限组管理、团队同步、发布触发等操作标签
- 新增操作颜色标识：删除为红色、创建为绿色、更新为黄色、其他为蓝色
- 新增审计日志 API：

+++ extensions/teax/server/api/orgs/index.get.ts
@@ -14,7 +14,7 @@ export default defineEventHandler(async (event) => {
     avatarUrl: schema.organizations.avatarUrl,
     syncedAt: schema.organizations.syncedAt,
     projectCount:
-      sql<number>`(SELECT COUNT(*) FROM projects WHERE projects.organization_id = "organizations"."id")`.as(
+      sql<number>`(SELECT COUNT(*)::int FROM projects WHERE projects.organization_id = "organizations"."id")`.as(
         "project_count",
       ),
   };
